Fix for `path` column header missing
Workaround for bug in tree-finder/regular-table. In certain conditions
(race condition), you can end up with two resize handles instead of the
assumed empty span + resize handle. (This might be due to memoization in
regular-table ?) In this case, we recover back to the state that
tree-finder was expecting. Can be removed once underlying issue is
resolved.