Fix rename bug + refresh
Fixes an issue where renaming a folder, would leave an entry that when
collapsed / expanded changed to a folder  named "/". the issue was that
Path.toarray expanded to a blank final element because of trailing
slash.

Also made sure that containing folder is refreshed after rename in order
to trigger a resort. There might be a more efficient way to only ask for a
resort without triggering a full request to the server.