Add load spinners
Adds some basic CSS spinners for when a content get is slow.
Currently we always add it, but it will be throttled by animation frames
/ draw requests to regular-table. Ideally we would have an initial wait,
but it would require managing timers. So there might be some flicker
when using a quick backend (but it will happen ~at the same time as
content comes in, so not too disruptive).

It is a little hackish, as we are tying into to contents proxy, instead
of the actual `expand` action in the grid, but that doesn't currently
have any hooks that we can use, so this is best effort.

The CSS spinners are simple in design since they are in the ::before
pseudo-element, so advanced structures like the JupyterLab Spinner
component are hard to reuse/replicate.